Refactor src/redux directory.
