+++ frontend/src/redux/reducers/index.js
@@ -1,16 +1,16 @@
 import { combineReducers } from "redux";
 import pdfViewer from "./pdf/pdfViewer/pdfViewer";
-import details from './ui/details/details'
-import skills from './ui/skills/skills'
-import hobbies from './ui/hobbies/hobbies'
-import experience from './ui/experience/experience'
-import languages from './ui/languages/languages'
-import education from './ui/education/education'
-import links from './ui/links/links'
-import gdpa from './ui/gdpa/gdpa'
-import certificates from './ui/certificates/certificates'
-import projects from './ui/projects/projects'
-import accessibility from './ui/accessibility/accessibility'
+import details from './ui/createcv/details/details'
+import skills from './ui/createcv/skills/skills'
+import hobbies from './ui/createcv/hobbies/hobbies'
+import experience from './ui/createcv/experience/experience'
+import languages from './ui/createcv/languages/languages'
+import education from './ui/createcv/education/education'
+import links from './ui/createcv/links/links'
+import gdpa from './ui/createcv/gdpa/gdpa'
+import certificates from './ui/createcv/certificates/certificates'
+import projects from './ui/createcv/projects/projects'
+import accessibility from './ui/components/accessibility/accessibility'
