+++ frontend/src/redux/reducers/ui/components/accessibility/actionTypes.js
