+++ frontend/src/redux/reducers/ui/components/accessibility/actions.js
