+++ frontend/src/redux/reducers/ui/createcv/certificates/certificates.js
@@ -1,6 +1,6 @@
 import * as actions from './actionTypes'
-import { formatterDateFormat } from '../../../../utilities/variables'
-import { getValidatedDate } from '../../../../utilities/date'
+import { formatterDateFormat } from '../../../../../utilities/variables'
+import { getValidatedDate } from '../../../../../utilities/date'
 import moment from 'moment'
 
 const initialState = {
