+++ frontend/src/redux/reducers/ui/createcv/details/actionTypes.js
