+++ frontend/src/redux/reducers/ui/createcv/details/details.js
