+++ frontend/src/redux/reducers/ui/createcv/education/education.js
@@ -1,6 +1,6 @@
 import * as actions from './actionTypes'
-import { getValidatedDate } from '../../../../utilities/date'
-import { formatterDateFormat } from '../../../../utilities/variables'
+import { getValidatedDate } from '../../../../../utilities/date'
+import { formatterDateFormat } from '../../../../../utilities/variables'
 import moment from 'moment'
 
 const initialState = {
