+++ frontend/src/redux/reducers/ui/createcv/experience/experience.js
@@ -1,6 +1,6 @@
 import * as actions from './actionTypes'
-import { getShortDateString, getValidatedDate } from '../../../../utilities/date'
-import {formatterDateFormat, } from '../../../../utilities/variables'
+import { getShortDateString, getValidatedDate } from '../../../../../utilities/date'
+import {formatterDateFormat, } from '../../../../../utilities/variables'
 import moment from 'moment'
 
 const initialState = {
