+++ frontend/src/redux/reducers/ui/createcv/links/actionTypes.js
