+++ frontend/src/redux/reducers/ui/createcv/projects/actionTypes.js
